+++ edgy/core/db/models/metaclasses.py
@@ -784,6 +784,8 @@ class BaseModelMeta(ModelMetaclass, ABCMeta):
         2. If a db_schema in the `registry` is passed, then it will use that as a default.
         3. If none is passed, defaults to the shared schema of the database connected.
         """
+        if cls.__is_proxy_model__:
+            return cls.__parent__.table  # type: ignore
         if not cls.meta.registry:
             # we cannot set the table without a registry
             # raising is required
@@ -841,6 +843,10 @@ class BaseModelMeta(ModelMetaclass, ABCMeta):
         Retrieve table for schema (nearly the same as build with scheme argument).
         Cache per class via a primitive LRU cache.
         """
+        if cls.__is_proxy_model__:
+            return cls.__parent__.table_schema(  # type: ignore
+                schema, metadata=metadata, update_cache=update_cache
+            )
         if schema is None or (cls.get_db_schema() or "") == schema:
             # sqlalchemy uses "" for empty schema
             table = getattr(cls, "_table", None)