+++ tests/clauses/test_and_clauses.py
@@ -78,18 +78,37 @@ async def test_filter_with_and_two():
     assert len(results) == 0
 
 
-async def test_filter_with_and_two_kwargs():
+async def test_filter_with_and_two_kwargs_with_user():
     user = await User.query.create(name="Adam", email="edgy@edgy.dev")
     await User.query.create(name="Edgy", email="edgy2@edgy.dev")
 
-    results = await User.query.filter(
-        and_.from_kwargs(User.columns, name="Adam", email="edgy@edgy.dev")
-    )
+    with pytest.warns(DeprecationWarning):
+        results = await User.query.filter(
+            and_.from_kwargs(User, name="Adam", email="edgy@edgy.dev")
+        )
 
     assert len(results) == 1
     assert results[0].pk == user.pk
 
-    results = await User.query.filter(and_.from_kwargs(User, name="Adam", email="edgy@edgy.dev"))
+    with pytest.warns(DeprecationWarning):
+        results = await User.query.filter(
+            and_.from_kwargs(User, name="Adam", email="edgy@edgy.dev")
+        )
+
+    assert len(results) == 1
+    assert results[0].pk == user.pk
+
+
+async def test_filter_with_and_two_kwargs_no_user():
+    user = await User.query.create(name="Adam", email="edgy@edgy.dev")
+    await User.query.create(name="Edgy", email="edgy2@edgy.dev")
+
+    results = await User.query.filter(and_.from_kwargs(name="Adam", email="edgy@edgy.dev"))
+
+    assert len(results) == 1
+    assert results[0].pk == user.pk
+
+    results = await User.query.filter(and_.from_kwargs(name="Adam", email="edgy@edgy.dev"))
 
     assert len(results) == 1
     assert results[0].pk == user.pk
@@ -159,6 +178,22 @@ async def test_filter_and_clause_related():
     assert results[0].pk == product.pk
 
     results = await User.query.and_(products__id=product.pk)
+
+    assert len(results) == 1
+    assert results[0].pk == user.pk
+
+
+async def test_filter_and_clause_related_helper():
+    user = await User.query.create(name="Adam", email="adam@edgy.dev")
+    await User.query.create(name="Edgy", email="edgy@edgy.dev")
+    product = await Product.query.create(user=user)
+
+    results = await Product.query.filter(and_.from_kwargs(user__id=user.pk))
+
+    assert len(results) == 1
+    assert results[0].pk == product.pk
+
+    results = await User.query.filter(and_.from_kwargs(products__id=product.pk))
 
     assert len(results) == 1
     assert results[0].pk == user.pk